input: move glfw parts in wato window class
